TXRX_LOOPBACK_TEST: correcting datatype ussed during recv to match cpu format for the rx streamer

+++ host/examples/txrx_trigger_loopback.cpp
@@ -76,10 +76,10 @@ void rx_run(uhd::rx_streamer::sptr rx_stream, double start_time, uint64_t num_tr
 
     // Vector containing each buffer
     // The buffer must be able to hold a full buffer's worth of data plus data from one more packet
-    std::vector<std::vector<std::complex<float>>> buffs(channel_nums.size(), std::vector<std::complex<float>>(samples_per_trigger*2));
+    std::vector<std::vector<std::complex<short>>> buffs(channel_nums.size(), std::vector<std::complex<short>>(samples_per_trigger*2));
 
     // Vector pointing to each buffer
-    std::vector<std::complex<float>*> buff_ptrs;
+    std::vector<std::complex<short>*> buff_ptrs;
     for (size_t i = 0; i < buffs.size(); i++)
         buff_ptrs.push_back(&buffs[i].front());
